==> notebooks/3.Cleaning_Restaurant_Matches.ipynb <==
import pickle
import pandas as pd
# ^^^ pyforest auto-imports - don't write above this line
# Imports

import unidecode
import wikipedia
import warnings
warnings.filterwarnings('ignore')

original_rest_matches_df = pd.read_csv("../my_saved_data/All_restaurant_matches_df.csv", \
                                       index_col=0).reset_index(drop=True)

with open("../my_saved_data/demonym_dictionary.pickle", "rb+") as f:
    demonym_dictionary = pickle.load(f)

## Removing any mistaken entries

with open("../my_saved_data/incorrect_matches.pickle", "rb+") as f:
    incorrect_matches = pickle.load(f)

rest_matches_df = original_rest_matches_df[~original_rest_matches_df['Matched_word'].isin(incorrect_matches)\
                                          ].reset_index(drop=True)

# Checking the matched word individually

mult_regions_df = rest_matches_df[rest_matches_df['Multiple_regions_flag'] == True]
problematic_matched_words = list(mult_regions_df.Matched_word.unique())

len(problematic_matched_words)

# mult_regions_df[mult_regions_df['Matched_word'] == 'leon']

# problematic_matched_words

single_regions_df = rest_matches_df[rest_matches_df['Multiple_regions_flag'] == False]
possible_problematic_matched_words = list(set(list(single_regions_df.Matched_word.unique())))

# how many regions are found in my study?
single_regions_df['Region'].nunique()

# single_regions_df[single_regions_df['Matched_word'] == 'bajio']

# possible_problematic_matched_words

# Using Wikipedia to correct any errors in data

# getting all of the multi-matched words in a single list 
multi_region_matches = list(set(rest_matches_df[rest_matches_df['Multiple_regions_flag']]['Matched_word']))
len(multi_region_matches)

# getting all of the state names 
state_names = list(demonym_dictionary.keys())

# replacement dictionary
replacement_dict = {key: None for key in rest_matches_df['Matched_word']}

rest_matches_df['Region'][6].count(",")

# adding words that are clearly tied to a region 
for idx, word in enumerate(rest_matches_df['Matched_word']):
    if rest_matches_df['Region'][idx].count(",") == 0:
        replacement_dict[word] = rest_matches_df['Region'][idx]

## Getting summaries from wikipedia and sorting results

summaries = []
somewhat_ambiguous_entries = []
ambiguous_entries = []
not_in_wikipedia = []
for idx, match in enumerate(multi_region_matches):
    try:
        summary = wikipedia.summary(match)
        str_match = " " + match + " "
        if str_match in summary:
            summaries.append(summary)
        elif str_match not in summary:
            summaries.append(idx)
            somewhat_ambiguous_entries.append(idx)
    except wikipedia.exceptions.DisambiguationError as e:
        if match + " (state)" in e.options:
            summary = wikipedia.summary(match + " (state)")
            summaries.append(summary)
        else:
            summaries.append(idx)
            ambiguous_entries.append((idx, e.options))
    except wikipedia.exceptions.PageError as p:
        not_in_wikipedia.append((idx, match)) 

# not_in_wikipedia

# replacing the not_in_wikipedia entry 
replacement_dict['lerma'] = 'Mexico'
replacement_dict['chalmita'] = 'Mexico'

# how many entries are still None?
none_counter = 0
for key, value in replacement_dict.items():
    if value == None:
        none_counter += 1
none_counter

## Adding entries that were found in wikipedia to dictionary

# how many summaries were found by wikipedia?
summaries_count = [x for x in summaries if type(x) != int]
print(f"wikipedia found {len(summaries_count)} entries")

summaries_dict = {'corralito' : None,
                 'copal': None,
                 'poblano': 'Puebla',
                 'guerrerense': "Guerrero",
                 'candela': None,
                 'leon': 'Nuevo Leon',
                 'morelia': "Puebla",
                 'coatzingo': None}

# for idx, x in enumerate(summaries):
#     if type(x) == str:
#         print(idx)
#         print(f"The entry is {multi_region_matches[idx]}")
#         print(x)

replacement_dict.update(summaries_dict)

# how many entries are still None?
none_counter = 0
for key, value in replacement_dict.items():
    if value == None:
        none_counter += 1
none_counter

## Googling Remaining Entries Without a Region

google_search_entries = {}
for key, value in replacement_dict.items():
    if value:
        continue
    if key in summaries_dict.keys():
        continue
    google_search_entries[key] = None

# carmen cooould refer to playa del carmen, but it clearly doesnt here 
google_search_entries = {'juarez': "Chihuahua",
                        'chinantla': 'Puebla',
                        'xochimilco': 'Mexico City',
                        'coyoacan': 'Mexico City',
                        'cholula': 'Puebla',
                        'matamoros': 'Tamaulipas',
                        'acatlan': 'Hidalgo',
                        'ocotlan': "Jalisco",
                        'atotonilco': 'Guanajuato',
                        'bustamante': 'Nuevo Leon',
                        'ayutla': "Jalisco",
                        'arandas': 'Jalisco',
                        'maravatio': 'Michoacan',
                        'uruapan': 'Michoacan',
                        'cocula': "Michoacan",
                        'camargo': "Chihuahua",
                        'cuernavaca': "Morelos",
                        'polanco': 'Mexico City',
                        'taxco': "Guerrero",
                         'zaragoza': 'Coahuila',
                         'monterrey': 'Nuevo Leon',
                         'tenochtitlan': 'Mexico City',
                         'aquila': "Michoacan",
                         'tuxpan': 'Veracruz',
                         'toluca': 'Mexico',
                         'ixtapa': 'Guerrero',
                         'apatzingan': 'Michoacan',
                         'quiroga': 'Michoacan',
                         'tuzantla': 'Michoacan',
                         'ahualulco': 'San Luis Potosi',
                         'picacho': 'Guanajuato',
                         'masatlan': 'Sinaloa'}
replacement_dict.update(google_search_entries)

## Replacing/Deleting Entries from DataFrame

# searching for incorrect region names 
for key, value in replacement_dict.items():
    if value != None:
        if value not in state_names:
            print(f"{key} has {value} spelt incorrectly")

rest_matches_df

# removing incorrect matched_words 
entries_to_remove = [key for key, value in replacement_dict.items() if value == None]
final_rest_matches_df = rest_matches_df[~rest_matches_df['Matched_word'].isin(entries_to_remove)].reset_index(drop=True)
# removed 31 entries
final_rest_matches_df.shape

final_rest_matches_df

# replacing the regions 
for idx, word in enumerate(final_rest_matches_df['Matched_word']):
    if word in replacement_dict.keys():
        final_rest_matches_df['Final_region'][idx] = replacement_dict[word]

final_rest_matches_df.head(15)

## Dropping columns

final_rest_matches_df.drop(columns = ['Region', 'Multiple_regions_flag'], inplace=True)

## Saving all_correct_restaurants 

final_rest_matches_df.to_csv("../my_saved_data/all_correct_restaurants_and_regions.csv")
